Remove front bundle
