+++ config/bundles.php
@@ -42,5 +42,4 @@ return [
     Vich\UploaderBundle\VichUploaderBundle::class => ['all' => true],
     Monofony\Bundle\CoreBundle\MonofonyCoreBundle::class => ['all' => true],
     Monofony\Bundle\AdminBundle\MonofonyAdminBundle::class => ['all' => true],
-    Monofony\Bundle\FrontBundle\MonofonyFrontBundle::class => ['all' => true],
 ];
