--- src/Monofony/Bundle/FrontBundle/DependencyInjection/MonofonyFrontBundleExtension.php
@@ -1,33 +1,0 @@
-<?php
-
-/*
- * This file is part of the Monofony package.
- *
- * (c) Monofony
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Monofony\Bundle\FrontBundle\DependencyInjection;
-
-use Monofony\Bundle\FrontBundle\Menu\AccountMenuBuilderInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\Extension;
-
-class MonofonyFrontBundleExtension extends Extension
-{
-    public function load(array $configs, ContainerBuilder $container): void
-    {
-        $this->buildAccountMenu($container);
-    }
-
-    private function buildAccountMenu(ContainerBuilder $container): void
-    {
-        $container->registerForAutoconfiguration(AccountMenuBuilderInterface::class)
-            ->addTag('knp_menu.menu_builder', [
-                'method' => 'createMenu',
-                'alias' => 'app.account',
-            ]);
-    }
-}
